pyproject.toml
[project]
name = "rag-agent-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.135.3",
    "langchain==1.2.15",
    "langchain-community==0.4.1",
    "langchain-core==1.2.29",
    "langchain-google-genai==4.2.2",
    "langchain-text-splitters==1.1.1",
    "pydantic==2.13.1",
    "python-dotenv==1.2.2",
]
